hover code cleanup
